feat(scheduler): add event handlers
