+++ scheduler/src/event/handlers/mod.rs
@@ -1,0 +1,6 @@
+pub mod instance_create;
+pub mod instance_destroy;
+pub mod instance_stop;
+pub mod node_register;
+pub mod node_status;
+pub mod node_unregister;
