+++ scheduler/src/event/mod.rs
@@ -1,0 +1,41 @@
+use std::net::IpAddr;
+
+use proto::scheduler::{
+    Instance, InstanceStatus, NodeRegisterRequest, NodeRegisterResponse, NodeStatus,
+    NodeUnregisterRequest, NodeUnregisterResponse,
+};
+use tokio::sync::{mpsc, oneshot};
+use tonic::Response;
+
+use crate::NodeIdentifier;
+
+pub mod handlers;
+
+#[derive(Debug)]
+pub enum Event {
+    // Instance events
+    InstanceCreate(
+        Instance,
+        mpsc::Sender<Result<InstanceStatus, tonic::Status>>,
+    ),
+    InstanceStop(
+        NodeIdentifier,
+        oneshot::Sender<Result<Response<()>, tonic::Status>>,
+    ),
+    InstanceDestroy(
+        NodeIdentifier,
+        oneshot::Sender<Result<Response<()>, tonic::Status>>,
+    ),
+
+    // Node events
+    NodeRegister(
+        NodeRegisterRequest,
+        IpAddr,
+        oneshot::Sender<Result<Response<NodeRegisterResponse>, tonic::Status>>,
+    ),
+    NodeUnregister(
+        NodeUnregisterRequest,
+        oneshot::Sender<Result<Response<NodeUnregisterResponse>, tonic::Status>>,
+    ),
+    NodeStatus(NodeStatus, mpsc::Sender<Result<(), tonic::Status>>),
+}
