+++ scheduler/src/lib.rs
@@ -1,17 +1,9 @@
-use std::{io, net::IpAddr};
-
-use proto::scheduler::{
-    Instance, InstanceStatus, NodeRegisterRequest, NodeRegisterResponse, NodeStatus,
-    NodeUnregisterRequest, NodeUnregisterResponse,
-};
+use std::io;
 use thiserror::Error;
-use tokio::{
-    sync::{mpsc, oneshot},
-    task::JoinError,
-};
-use tonic::Response;
+use tokio::task::JoinError;
 
 pub mod config;
+pub mod event;
 pub mod instance;
 pub mod manager;
 pub mod node;
@@ -59,32 +51,3 @@ pub enum ProxyError {
 
 pub type NodeIdentifier = String;
 pub type InstanceIdentifier = String;
-
-#[derive(Debug)]
-pub enum Event {
-    // Instance events
-    InstanceCreate(
-        Instance,
-        mpsc::Sender<Result<InstanceStatus, tonic::Status>>,
-    ),
-    InstanceStop(
-        NodeIdentifier,
-        oneshot::Sender<Result<Response<()>, tonic::Status>>,
-    ),
-    InstanceDestroy(
-        NodeIdentifier,
-        oneshot::Sender<Result<Response<()>, tonic::Status>>,
-    ),
-
-    // Node events
-    NodeRegister(
-        NodeRegisterRequest,
-        IpAddr,
-        oneshot::Sender<Result<Response<NodeRegisterResponse>, tonic::Status>>,
-    ),
-    NodeUnregister(
-        NodeUnregisterRequest,
-        oneshot::Sender<Result<Response<NodeUnregisterResponse>, tonic::Status>>,
-    ),
-    NodeStatus(NodeStatus, mpsc::Sender<Result<(), tonic::Status>>),
-}
